Updated some formatting, and fixed an issue with single-inputs not having  the proper name id when dynamically added to the html template.
For single inputs, the naming convention used in "addInput" in formatters,
which is  responsible for dynamically adding submittable input to the html
document, was declaring the name as "input1" for all  uses. This was fine
when it was only being used for the first single input, but prevented reuse.
Refactoring earlier would have been smart, and allowed for a simpler
multi-input method.